Perform Brain Gene Set Analysis

This function performs a gene set analysis using brain data.

Usage

brainenrich(
  brain_data,
  gene_data,
  annoData,
  cor_method = c("pearson", "spearman", "pls1c", "pls1w", "custom"),
  aggre_method = c("mean", "median", "meanabs", "meansqr", "maxmean", "ks_orig",
    "ks_weighted", "ks_pos_neg_sum", "local_fdr", "sign_test", "rank_sum", "custom"),
  null_model = c("spin_brain", "resample_gene", "coexp_matched"),
  n_perm = 5000,
  perm_id = NULL,
  coord.l = NULL,
  coord.r = NULL,
  seed = NULL,
  n_cores = 0,
  minGSSize = 10,
  maxGSSize = 200,
  thres_type = c("sd", "percentile"),
  thres_val = 1,
  pvalueCutoff = 0.05,
  pAdjustMethod = "fdr",
  matchcoexp_tol = 0.05,
  matchcoexp_max_iter = 1e+06
)

Arguments

brain_data

A data frame of brain data. Region by 1 column.

gene_data

A data frame of gene expression data.

annoData

An environment containing annotation data.

cor_method

A character string specifying the correlation method. Default is 'pearson'. Other options include 'spearman', 'pls1c', 'pls1w', 'custom'.

aggre_method

A character string specifying the aggregation method. Default is 'mean'. Other options include 'median', 'meanabs', 'meansqr', 'maxmean', 'ks_orig', 'ks_weighted', 'ks_pos_neg_sum', 'local_fdr', 'sign_test', 'rank_sum', 'custom'.

null_model

A character string specifying the null model. Default is 'spin_brain'. Other options include 'resample_gene', 'coexp_matched'.

n_perm

An integer specifying the number of permutations. Default is 5000.

perm_id

A matrix of permutation IDs. Default is NULL.

coord.l

A matrix of left hemisphere coordinates. Default is NULL.

coord.r

A matrix of right hemisphere coordinates. Default is NULL.

seed

An integer specifying the seed for reproducibility of spinning brain. Default is NULL.

n_cores

An integer specifying the number of cores to use. Default is 0.

minGSSize

An integer specifying the minimum gene set size. Default is 10.

maxGSSize

An integer specifying the maximum gene set size. Default is 200.

thres_type

A character string specifying the threshold type for core genes. Default is 'sd'. Other option is 'percentile'.

thres_val

A numeric value specifying the threshold value for core genes. Default is 1.

pvalueCutoff

A numeric value specifying the p-value cutoff for output. Default is 0.05.

pAdjustMethod

A character string specifying the method for p-value adjustment. Default is 'fdr'.

matchcoexp_tol

A numeric value specifying the tolerance for matched co-expression. Lower value means better matching but will take much more iterations. Default is 0.05.

matchcoexp_max_iter

An integer specifying the maximum number of iterations for matched co-expression. Default is 1000000.

Value

A gseaResult object containing the enrichment results.