Find Core Genes Influencing Aggregated Score or LM Coefficients between molecular profile and behavioral data

This function performs a Leave-One-Out (LOO) analysis on gene sets to determine core genes that influence the aggregated score. It can utilize parallel processing to enhance computation efficiency and supports two types of analysis: one that considers only gene sets and another that includes predictor and covariate data frames.

Usage

find_core_genes(
  geneList,
  geneSetList,
  pred_df = NULL,
  cov_df = NULL,
  aggre_method,
  n_cores = 1,
  threshold_type = c("sd", "percentile"),
  threshold_value = 1
)

Arguments

geneList

A matrix of genes by subs, each column representing a subject / a group-level result.

geneSetList

A list of gene sets, each containing names of genes.

pred_df

Optional data frame of a predictor. If NULL, it is perfomred for group-level enrichment.

cov_df

Optional data frame of covariates. If NULL, it is perfomred for group-level enrichment.

aggre_method

The aggregation method used to compute the scores.

n_cores

The number of cores to use for parallel processing; defaults to 1. Uses all available cores minus one if set to 0.

threshold_type

The method to determine significance ('sd' for standard deviation, 'percentile' for percentile threshold).

threshold_value

Numeric value specifying the threshold level; meaning depends on threshold_type.

Value

A list of core genes for each gene set.