Generate Null Brain Data

This function generates null brain datasets based on permutations provided in perm_id. It rearranges the brain data according to the permutations and outputs the shuffled datasets.

Usage

generate_null_brain_data(brain_data, perm_id)

Arguments

brain_data

A matrix representing brain data, where each row corresponds to a region.

perm_id

A matrix of permutations, where each column represents a permutation and each row corresponds to an index in brain_data.

Value

A matrix of null brain data with the same dimensions as brain_data but with permuted rows according to perm_id.