Load gene sets annotation data

This function loads annotation data from RDS files located in the inst/extdata/geneSets directory of the package. If the specified file does not exist locally, it will be downloaded from the GitHub repository.

Usage

get_annoData(
  type = c("CellTypes_Lake2018", "CellTypes_Martins2021", "CellTypes_Seidlitz2020",
    "DGN", "GO_BP", "GO_CC", "GO_MF", "KEGG", "Reactome", "SynGO", "WikiPathways",
    "MeSH_A", "MeSH_B", "MeSH_C", "MeSH_D", "MeSH_E", "MeSH_F", "MeSH_G", "MeSH_H",
    "MeSH_I", "MeSH_J", "MeSH_K", "MeSH_L", "MeSH_M", "MeSH_N", "MeSH_Z")
)

Arguments

type

A character string specifying the type of gene set to load. Options are:

"CellTypes_Lake2018"

Cell types data from Lake2018

"CellTypes_Martins2021"

Cell types data from Martins2021

"CellTypes_Seidlitz2020"

Cell types data from Seidlitz2020

"DGN"

DisGeNET gene sets

"GO_BP"

Gene Ontology Biological Process

"GO_CC"

Gene Ontology Cellular Component

"GO_MF"

Gene Ontology Molecular Function

"KEGG"

KEGG gene sets

"Reactome"

Reactome gene sets

"SynGO"

SynGO gene sets

"WikiPathways"

WikiPathways gene sets

"MeSH_A"

MeSH category A gene sets

"MeSH_B"

MeSH category B gene sets

"MeSH_C"

MeSH category C gene sets

"MeSH_D"

MeSH category D gene sets

"MeSH_E"

MeSH category E gene sets

"MeSH_F"

MeSH category F gene sets

"MeSH_G"

MeSH category G gene sets

"MeSH_H"

MeSH category H gene sets

"MeSH_I"

MeSH category I gene sets

"MeSH_J"

MeSH category J gene sets

"MeSH_K"

MeSH category K gene sets

"MeSH_L"

MeSH category L gene sets

"MeSH_M"

MeSH category M gene sets

"MeSH_N"

MeSH category N gene sets

"MeSH_Z"

MeSH category Z gene sets

Value

A data frame containing the annotation data.

Examples

if (FALSE) { # \dontrun{
annoData <- get_annoData("GO_BP")
} # }