Get Gene Expression Data

This function retrieves gene expression data based on specified parameters. #' This function loads gene expression data from CSV files located in the inst/extdata/geneExp directory of the package. If the specified file does not exist locally, it will be downloaded from the GitHub repository.

Usage

get_geneExp(
  atlas = c("desikan", "schaefer100", "schaefer200", "schaefer300"),
  rdonor = c("r0.2", "r0.4", "r0.6"),
  hem = c("L", "R", "B")
)

Arguments

atlas

A character string specifying the atlas to use. Options are "desikan", "schaefer100", "schaefer200", "schaefer300".

rdonor

A character string specifying the donor resolution to use. Options are "r0.2", "r0.4", "r0.6".

hem

A character string specifying the hemisphere to use. Options are "L" (Left), "R" (Right), "B" (Both).

Value

A matrix containing the gene expression data.

Details

The data is obtained from the ENIGMA-TOOLBOX. Please cite the ENIGMA-TOOLBOX: Larivière, S., Paquola, C., Park, B. Y., Royer, J., Wang, Y., Benkarim, O., ... & Bernhardt, B. C. (2021). The ENIGMA Toolbox: multiscale neural contextualization of multisite neuroimaging datasets. Nature Methods, 18(7), 698-700.

Examples

if (FALSE) { # \dontrun{
geneExpMatrix <- get_geneExp("desikan", "r0.4", "L")
} # }