Rotate Parcellation

Generate a permutation map from a set of cortical regions of interest to itself, while (approximately) preserving contiguity and hemispheric symmetry. The function is based on a rotation of the FreeSurfer projection of coordinates of a set of regions of interest on the sphere. #' This function is modified from the original version available at: https://github.com/frantisekvasa/rotate_parcellation

Usage

rotate_parcellation(
  coord.l = NULL,
  coord.r = NULL,
  nrot = 5000,
  method = c("hungarian", "vasa"),
  seed = NULL
)

Arguments

coord.l

Coordinates of left hemisphere regions on the sphere (array of size n(LH regions) x 3). Can be NULL if only right hemisphere is used.

coord.r

Coordinates of right hemisphere regions on the sphere (array of size n(RH regions) x 3). Can be NULL if only left hemisphere is used.

nrot

Number of rotations (default = 5000).

method

Method to match rotated and unrotated regions; options are 'vasa' (faster, can be suboptimal) or 'hungarian' (default, slower, optimal).

Value

Array of permutations, from set of regions to itself (array of size n(total regions) x nrot).

Details

Modifications include:

Added support for scenarios where only one hemisphere's coordinates are provided.

Improved handling of coordinate dimensions and conditional concatenation of reference and rotation indices.

Included importFrom directives for required functions from matrixStats and clue.

Ensured the function generates nrot + 100 permutations, removes duplicates, and returns exactly nrot unique permutations.

Examples

# Example usage with both hemispheres
coord.l <- matrix(runif(30), nrow = 10, ncol = 3)
coord.r <- matrix(runif(30), nrow = 10, ncol = 3)
permutations <- rotate_parcellation(coord.l, coord.r)
#> permutation 100 of 6000
#> map to itself n. 1
#> map to itself n. 2
#> map to itself n. 3
#> permutation 200 of 6000
#> map to itself n. 4
#> map to itself n. 5
#> permutation 300 of 6000
#> map to itself n. 6
#> permutation 400 of 6000
#> map to itself n. 7
#> permutation 500 of 6000
#> permutation 600 of 6000
#> map to itself n. 8
#> permutation 700 of 6000
#> permutation 800 of 6000
#> permutation 900 of 6000
#> map to itself n. 9
#> permutation 1000 of 6000
#> permutation 1100 of 6000
#> permutation 1200 of 6000
#> permutation 1300 of 6000
#> permutation 1400 of 6000
#> permutation 1500 of 6000
#> permutation 1600 of 6000
#> permutation 1700 of 6000
#> map to itself n. 10
#> permutation 1800 of 6000
#> map to itself n. 11
#> permutation 1900 of 6000
#> map to itself n. 12
#> map to itself n. 13
#> map to itself n. 14
#> permutation 2000 of 6000
#> permutation 2100 of 6000
#> map to itself n. 15
#> map to itself n. 16
#> permutation 2200 of 6000
#> permutation 2300 of 6000
#> map to itself n. 17
#> map to itself n. 18
#> permutation 2400 of 6000
#> map to itself n. 19
#> permutation 2500 of 6000
#> map to itself n. 20
#> permutation 2600 of 6000
#> map to itself n. 21
#> map to itself n. 22
#> permutation 2700 of 6000
#> map to itself n. 23
#> map to itself n. 24
#> permutation 2800 of 6000
#> map to itself n. 25
#> permutation 2900 of 6000
#> permutation 3000 of 6000
#> map to itself n. 26
#> map to itself n. 27
#> map to itself n. 28
#> permutation 3100 of 6000
#> map to itself n. 29
#> permutation 3200 of 6000
#> map to itself n. 30
#> permutation 3300 of 6000
#> map to itself n. 31
#> map to itself n. 32
#> map to itself n. 33
#> permutation 3400 of 6000
#> permutation 3500 of 6000
#> map to itself n. 34
#> permutation 3600 of 6000
#> map to itself n. 35
#> permutation 3700 of 6000
#> map to itself n. 36
#> permutation 3800 of 6000
#> permutation 3900 of 6000
#> map to itself n. 37
#> permutation 4000 of 6000
#> map to itself n. 38
#> permutation 4100 of 6000
#> map to itself n. 39
#> permutation 4200 of 6000
#> map to itself n. 40
#> permutation 4200 of 6000
#> map to itself n. 41
#> permutation 4300 of 6000
#> map to itself n. 42
#> map to itself n. 43
#> map to itself n. 44
#> permutation 4400 of 6000
#> map to itself n. 45
#> permutation 4500 of 6000
#> map to itself n. 46
#> permutation 4600 of 6000
#> map to itself n. 47
#> permutation 4700 of 6000
#> map to itself n. 48
#> permutation 4800 of 6000
#> permutation 4900 of 6000
#> map to itself n. 49
#> map to itself n. 50
#> permutation 5000 of 6000
#> map to itself n. 51
#> permutation 5100 of 6000
#> map to itself n. 52
#> map to itself n. 53
#> map to itself n. 54
#> permutation 5200 of 6000
#> map to itself n. 55
#> permutation 5300 of 6000
#> map to itself n. 56
#> permutation 5400 of 6000
#> map to itself n. 57
#> permutation 5500 of 6000
#> permutation 5600 of 6000
#> map to itself n. 58
#> map to itself n. 59
#> permutation 5700 of 6000
#> map to itself n. 60
#> permutation 5800 of 6000
#> map to itself n. 61
#> permutation 5900 of 6000
#> permutation 6000 of 6000

# Example usage with one hemisphere (left)
coord.l <- matrix(runif(30), nrow = 10, ncol = 3)
permutations <- rotate_parcellation(coord.l, NULL)
#> map to itself n. 1
#> permutation 100 of 6000
#> map to itself n. 2
#> map to itself n. 3
#> map to itself n. 4
#> permutation 200 of 6000
#> map to itself n. 5
#> map to itself n. 6
#> map to itself n. 7
#> map to itself n. 8
#> permutation 300 of 6000
#> map to itself n. 9
#> map to itself n. 10
#> map to itself n. 11
#> permutation 400 of 6000
#> map to itself n. 12
#> map to itself n. 13
#> map to itself n. 14
#> map to itself n. 15
#> map to itself n. 16
#> permutation 500 of 6000
#> map to itself n. 17
#> map to itself n. 18
#> permutation 600 of 6000
#> map to itself n. 19
#> map to itself n. 20
#> map to itself n. 21
#> map to itself n. 22
#> map to itself n. 23
#> permutation 700 of 6000
#> map to itself n. 24
#> map to itself n. 25
#> permutation 800 of 6000
#> map to itself n. 26
#> map to itself n. 27
#> map to itself n. 28
#> map to itself n. 29
#> map to itself n. 30
#> permutation 900 of 6000
#> map to itself n. 31
#> map to itself n. 32
#> permutation 1000 of 6000
#> map to itself n. 33
#> map to itself n. 34
#> permutation 1100 of 6000
#> permutation 1200 of 6000
#> map to itself n. 35
#> map to itself n. 36
#> map to itself n. 37
#> map to itself n. 38
#> map to itself n. 39
#> map to itself n. 40
#> permutation 1300 of 6000
#> map to itself n. 41
#> map to itself n. 42
#> permutation 1400 of 6000
#> map to itself n. 43
#> map to itself n. 44
#> map to itself n. 45
#> permutation 1500 of 6000
#> map to itself n. 46
#> map to itself n. 47
#> permutation 1600 of 6000
#> map to itself n. 48
#> map to itself n. 49
#> permutation 1700 of 6000
#> map to itself n. 50
#> map to itself n. 51
#> map to itself n. 52
#> permutation 1800 of 6000
#> map to itself n. 53
#> map to itself n. 54
#> map to itself n. 55
#> map to itself n. 56
#> map to itself n. 57
#> permutation 1900 of 6000
#> map to itself n. 58
#> map to itself n. 59
#> map to itself n. 60
#> permutation 2000 of 6000
#> map to itself n. 61
#> permutation 2100 of 6000
#> map to itself n. 62
#> permutation 2200 of 6000
#> map to itself n. 63
#> map to itself n. 64
#> map to itself n. 65
#> permutation 2300 of 6000
#> map to itself n. 66
#> permutation 2400 of 6000
#> map to itself n. 67
#> map to itself n. 68
#> map to itself n. 69
#> map to itself n. 70
#> permutation 2500 of 6000
#> map to itself n. 71
#> permutation 2600 of 6000
#> map to itself n. 72
#> permutation 2700 of 6000
#> map to itself n. 73
#> map to itself n. 74
#> map to itself n. 75
#> permutation 2800 of 6000
#> map to itself n. 76
#> map to itself n. 77
#> map to itself n. 78
#> map to itself n. 79
#> permutation 2900 of 6000
#> map to itself n. 80
#> permutation 3000 of 6000
#> permutation 3100 of 6000
#> map to itself n. 81
#> map to itself n. 82
#> map to itself n. 83
#> map to itself n. 84
#> map to itself n. 85
#> permutation 3200 of 6000
#> map to itself n. 86
#> permutation 3300 of 6000
#> map to itself n. 87
#> permutation 3400 of 6000
#> permutation 3500 of 6000
#> map to itself n. 88
#> map to itself n. 89
#> map to itself n. 90
#> permutation 3600 of 6000
#> map to itself n. 91
#> permutation 3700 of 6000
#> map to itself n. 92
#> map to itself n. 93
#> map to itself n. 94
#> map to itself n. 95
#> permutation 3800 of 6000
#> map to itself n. 96
#> permutation 3900 of 6000
#> map to itself n. 97
#> permutation 4000 of 6000
#> map to itself n. 98
#> map to itself n. 99
#> map to itself n. 100
#> permutation 4100 of 6000
#> map to itself n. 101
#> map to itself n. 102
#> map to itself n. 103
#> map to itself n. 104
#> permutation 4200 of 6000
#> permutation 4300 of 6000
#> map to itself n. 105
#> permutation 4400 of 6000
#> map to itself n. 106
#> map to itself n. 107
#> permutation 4500 of 6000
#> map to itself n. 108
#> permutation 4600 of 6000
#> map to itself n. 109
#> permutation 4700 of 6000
#> map to itself n. 110
#> map to itself n. 111
#> permutation 4800 of 6000
#> map to itself n. 112
#> map to itself n. 113
#> permutation 4900 of 6000
#> map to itself n. 114
#> map to itself n. 115
#> permutation 5000 of 6000
#> map to itself n. 116
#> map to itself n. 117
#> map to itself n. 118
#> permutation 5100 of 6000
#> map to itself n. 119
#> map to itself n. 120
#> permutation 5200 of 6000
#> permutation 5300 of 6000
#> map to itself n. 121
#> permutation 5400 of 6000
#> map to itself n. 122
#> map to itself n. 123
#> permutation 5500 of 6000
#> map to itself n. 124
#> map to itself n. 125
#> permutation 5600 of 6000
#> map to itself n. 126
#> map to itself n. 127
#> map to itself n. 128
#> map to itself n. 129
#> permutation 5700 of 6000
#> map to itself n. 130
#> map to itself n. 131
#> map to itself n. 132
#> permutation 5800 of 6000
#> map to itself n. 133
#> map to itself n. 134
#> permutation 5900 of 6000
#> map to itself n. 135
#> permutation 6000 of 6000

# Example usage with one hemisphere (right)
coord.r <- matrix(runif(30), nrow = 10, ncol = 3)
permutations <- rotate_parcellation(NULL, coord.r)
#> map to itself n. 1
#> permutation 100 of 6000
#> map to itself n. 2
#> map to itself n. 3
#> map to itself n. 4
#> permutation 200 of 6000
#> map to itself n. 5
#> permutation 300 of 6000
#> map to itself n. 6
#> map to itself n. 7
#> map to itself n. 8
#> permutation 400 of 6000
#> map to itself n. 9
#> permutation 500 of 6000
#> map to itself n. 10
#> permutation 600 of 6000
#> map to itself n. 11
#> permutation 600 of 6000
#> map to itself n. 12
#> map to itself n. 13
#> map to itself n. 14
#> permutation 700 of 6000
#> map to itself n. 15
#> permutation 800 of 6000
#> map to itself n. 16
#> permutation 900 of 6000
#> permutation 1000 of 6000
#> map to itself n. 17
#> permutation 1100 of 6000
#> map to itself n. 18
#> permutation 1200 of 6000
#> map to itself n. 19
#> map to itself n. 20
#> map to itself n. 21
#> permutation 1300 of 6000
#> permutation 1400 of 6000
#> map to itself n. 22
#> map to itself n. 23
#> map to itself n. 24
#> map to itself n. 25
#> permutation 1500 of 6000
#> map to itself n. 26
#> permutation 1600 of 6000
#> permutation 1700 of 6000
#> map to itself n. 27
#> permutation 1800 of 6000
#> permutation 1900 of 6000
#> map to itself n. 28
#> map to itself n. 29
#> permutation 2000 of 6000
#> permutation 2100 of 6000
#> map to itself n. 30
#> permutation 2200 of 6000
#> map to itself n. 31
#> map to itself n. 32
#> map to itself n. 33
#> permutation 2300 of 6000
#> map to itself n. 34
#> map to itself n. 35
#> map to itself n. 36
#> map to itself n. 37
#> permutation 2400 of 6000
#> map to itself n. 38
#> map to itself n. 39
#> map to itself n. 40
#> map to itself n. 41
#> permutation 2500 of 6000
#> map to itself n. 42
#> permutation 2600 of 6000
#> map to itself n. 43
#> map to itself n. 44
#> map to itself n. 45
#> map to itself n. 46
#> permutation 2700 of 6000
#> map to itself n. 47
#> map to itself n. 48
#> map to itself n. 49
#> permutation 2800 of 6000
#> map to itself n. 50
#> map to itself n. 51
#> permutation 2900 of 6000
#> map to itself n. 52
#> map to itself n. 53
#> permutation 3000 of 6000
#> map to itself n. 54
#> permutation 3100 of 6000
#> map to itself n. 55
#> permutation 3200 of 6000
#> permutation 3300 of 6000
#> map to itself n. 56
#> map to itself n. 57
#> permutation 3400 of 6000
#> map to itself n. 58
#> map to itself n. 59
#> permutation 3500 of 6000
#> map to itself n. 60
#> map to itself n. 61
#> permutation 3600 of 6000
#> permutation 3700 of 6000
#> map to itself n. 62
#> map to itself n. 63
#> map to itself n. 64
#> permutation 3800 of 6000
#> map to itself n. 65
#> map to itself n. 66
#> permutation 3900 of 6000
#> map to itself n. 67
#> permutation 4000 of 6000
#> permutation 4100 of 6000
#> map to itself n. 68
#> map to itself n. 69
#> permutation 4200 of 6000
#> map to itself n. 70
#> map to itself n. 71
#> map to itself n. 72
#> map to itself n. 73
#> map to itself n. 74
#> map to itself n. 75
#> permutation 4300 of 6000
#> map to itself n. 76
#> map to itself n. 77
#> map to itself n. 78
#> permutation 4400 of 6000
#> map to itself n. 79
#> permutation 4500 of 6000
#> permutation 4600 of 6000
#> map to itself n. 80
#> permutation 4700 of 6000
#> map to itself n. 81
#> map to itself n. 82
#> permutation 4800 of 6000
#> map to itself n. 83
#> map to itself n. 84
#> permutation 4900 of 6000
#> permutation 5000 of 6000
#> map to itself n. 85
#> map to itself n. 86
#> permutation 5100 of 6000
#> map to itself n. 87
#> map to itself n. 88
#> map to itself n. 89
#> permutation 5200 of 6000
#> permutation 5300 of 6000
#> map to itself n. 90
#> map to itself n. 91
#> permutation 5400 of 6000
#> map to itself n. 92
#> map to itself n. 93
#> permutation 5500 of 6000
#> map to itself n. 94
#> permutation 5600 of 6000
#> map to itself n. 95
#> permutation 5700 of 6000
#> map to itself n. 96
#> map to itself n. 97
#> map to itself n. 98
#> permutation 5800 of 6000
#> map to itself n. 99
#> map to itself n. 100
#> permutation 5900 of 6000
#> map to itself n. 101
#> map to itself n. 102
#> permutation 6000 of 6000